move accept reject matching ligand to snack bar
